obs/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
python-multipart==0.0.9
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
